Split common utilities by functionality
