+++ src/cfa/stf/routine/dagster_defs.py
@@ -35,14 +35,14 @@ from pyrenew_multisignal.hew.utils import flags_from_hew_letters
 from cfa.stf.routine._paths import PRODUCTION_PRIORS
 from cfa.stf.routine.fable.forecast_fable import main as forecast_fable
 from cfa.stf.routine.pyrenew_hew.forecast_pyrenew import main as forecast_pyrenew
-from cfa.stf.routine.utils.common_utils import (
-    calculate_training_dates,
-    create_prop_samples,
-    get_model_batch_dir_name,
+from cfa.stf.routine.utils.date_utils import calculate_training_dates
+from cfa.stf.routine.utils.directory_utils import get_model_batch_dir_name
+from cfa.stf.routine.utils.postprocess_forecast_batches import main as postprocess
+from cfa.stf.routine.utils.prop_utils import create_prop_samples
+from cfa.stf.routine.utils.r_utils import (
     make_figures_from_model_fit_dir,
     model_fit_dir_to_hub_tbl,
 )
-from cfa.stf.routine.utils.postprocess_forecast_batches import main as postprocess
 
 # ============================================================================
 # DAGSTER INITIALIZATION
