+++ src/cfa/stf/routine/epiautogp/forecast_epiautogp.py
@@ -17,11 +17,9 @@ from cfa.stf.routine.epiautogp.prep_epiautogp_data import (
 from cfa.stf.routine.epiautogp.reporting_delay_nowcast import ReportingDelayNowcast
 from cfa.stf.routine.forecast_pipeline import ForecastPipeline
 from cfa.stf.routine.forecast_run import ForecastRun
-from cfa.stf.routine.utils.common_utils import (
-    generate_epiweekly_data,
-    parse_exclude_date_ranges,
-    run_julia_script,
-)
+from cfa.stf.routine.utils.data_utils import generate_epiweekly_data
+from cfa.stf.routine.utils.date_utils import parse_exclude_date_ranges
+from cfa.stf.routine.utils.language_utils import run_julia_script
 
 _FIT_SCRIPT = Path(__file__).parent / "fit_epiautogp.jl"
 NowcastSourceName = Literal["none", "reporting-delay", "hubverse"]
